% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_kelvin.R
\name{celsius_to_kelvin}
\alias{celsius_to_kelvin}
\title{convert a temperature from celcius to kelvin}
\usage{
celsius_to_kelvin(temp_c)
}
\arguments{
\item{temp_c}{numeric. A temperature measured in celsius}
}
\value{
numeric. A temperature measured in kelvin
}
\description{
convert a temperature from celcius to kelvin
}
\examples{
celsius_to_kelvin(0)
}
